+++ test.cpp
@@ -1,6 +1,7 @@
 #include "gmock/gmock.h"
 #include "kiwer_api.cpp"
 #include "nemo_api.cpp"
+#include "AutoTradingSystem.cpp"
 #include <string>
 
 using namespace testing;
@@ -17,73 +18,6 @@ using namespace std;
 
 // Interface정의 및 구현. 다른 파일로 이동하셔도 좋을 것 같습니다.
 
-class IStockBrockerDriver
-{
-public:
-	virtual ~IStockBrockerDriver() = default;
-	virtual void login(std::string id, std::string password) = 0;
-	virtual void buy(std::string stockCode, int price, int count) = 0;
-	virtual void sell(std::string stockCode, int price, int count) = 0;
-	virtual void getPrice(std::string stockCode) = 0;
-};
-
-class KiwiDriver : public IStockBrockerDriver
-{
-    void login(std::string id, std::string password)
-    {
-        Kiwer.login(id, password);
-    };
-    void buy(std::string stockCode, int price, int count){
-        Kiwer.buy(stockCode, count, price);
-    };
-    void sell(std::string stockCode, int price, int count){
-        Kiwer.sell(stockCode, price, count);
-    };
-    void getPrice(std::string stockCode){};
-private:
-    KiwerAPI Kiwer;
-};
-
-class NemoDriver : public IStockBrockerDriver
-{
-    void login(std::string id, std::string password)
-    {
-        Nemo.certification(id, password);
-    };
-    void buy(std::string stockCode, int price, int count){
-        Nemo.purchasingStock(stockCode, price, count);
-    };
-    void sell(std::string stockCode, int price, int count){
-        Nemo.sellingStock(stockCode, price, count);
-    };
-    void getPrice(std::string stockCode){};
-private:
-    NemoAPI Nemo;
-};
-
-class AutoTradingSystem {
-public:
-    int selectStockBroker(string brokerName) {
-        int ret = 0;
-
-        if (brokerName == "KIWER") {
-            StockBrockerDriver = new KiwiDriver;
-        }
-        else if (brokerName == "NEMO") {
-            StockBrockerDriver = new NemoDriver;
-        }
-        else {
-            ret = -1;
-        }
-        return ret;
-    }
-
-    IStockBrockerDriver* getStockBroker() {
-        return StockBrockerDriver;
-    }
-   
-    IStockBrockerDriver *StockBrockerDriver = nullptr;
-};
 
 // Test용 Fixture
 class StockBrokerDriverTest : public::testing::Test {
@@ -100,7 +34,7 @@ protected:
 class KiwiDriverTest : public StockBrokerDriverTest {
 protected:
     std::unique_ptr<IStockBrockerDriver> createDriver() override {
-        return std::make_unique<KiwiDriver>();
+        return std::make_unique<KiwerDriver>();
     }
 };
